programs/password_manager/2-CoT-haiku4.5/requirements.txt
Flask>=3.0
Flask-Login>=0.6
Flask-WTF>=1.2
Flask-SQLAlchemy>=3.1
Werkzeug>=3.0
cryptography>=42.0
python-dotenv>=1.0
email-validator>=2.1
